fix(scmi-msg): base: fix build break
fix below build error due to strict type checking

include/lib/utils_def.h:70:20: error: comparison of distinct pointer types lacks a cast [-Werror]
   70 |         (void)(&_x == &_y);             \
      |                    ^~
drivers/scmi-msg/base.c:154:25: note: in expansion of macro ‘MIN’
  154 |                 count = MIN(count - a2p->skip, msg->out_size - sizeof(p2a));
      |                         ^~~

+++ drivers/scmi-msg/base.c
@@ -151,7 +151,7 @@ static void discover_list_protocols(struct scmi_msg *msg)
 	count = count_protocols_in_list(list);
 
 	if (count > a2p->skip) {
-		count = MIN(count - a2p->skip, msg->out_size - sizeof(p2a));
+		count = MIN(count - a2p->skip, (unsigned int)(msg->out_size - sizeof(p2a)));
 	} else {
 		count = 0U;
 	}
